* src/java/org/archive/crawler/datamodel/CandidateURI.java     (getInt): Was returning a long instead of an int (Kris pointed this out).

+++ heritrix/src/java/org/archive/crawler/datamodel/CandidateURI.java
@@ -443,7 +443,7 @@ implements Serializable, Lineable {
         getAList().putLong(key, value);
     }
     
-    public long getInt(String key) {
+    public int getInt(String key) {
         return getAList().getInt(key);
     }
